+++ src/DisableBigImageSizeTreshold.php
@@ -1,0 +1,20 @@
+<?php
+/**
+ * Disable big image size treshold.
+ */
+
+namespace Geniem\Project;
+
+/**
+ * DisableBigImageSizeTreshold class.
+ */
+class DisableBigImageSizeTreshold {
+
+    /**
+     * The constructor.
+     */
+    public function __construct() {
+
+        \add_filter( 'big_image_size_threshold', '__return_false' );
+    }
+}